requirements.txt
azure-identity
bioc==2.0.post4
datasets>=2.8.0,<3.0.0
evaluate
jinja2>=3.1.4
jsonschema
langchain
langchain-community
langchain-openai==0.1.14
nltk==3.8.1
numpy>=1.21.2,<2.0.0
openai>=1.0.0
pandas>=1.3.3
pydantic<3
python-dotenv
requests
scikit-learn
scipy>=1.9.0
sentencepiece
tiktoken
tqdm
transformers>=4.34.0
